+++ library/LibUtilities/TimeIntegration/IMEXdirkTimeIntegrationSchemes.h
@@ -278,8 +278,8 @@ public:
     LUE static void SetupSchemeData_2_2_2(
         TimeIntegrationAlgorithmGLMSharedPtr &phase)
     {
-        NekDouble glambda = 0.2928932188134524756;
-        NekDouble gdelta  = -0.7071067811865475244;
+        NekDouble glambda = 1.0 - sqrt(2.0) / 2.0;
+        NekDouble gdelta  = -sqrt(2.0) / 2.0;
 
         phase->m_A[0][1][1] = glambda;
         phase->m_A[0][2][1] = 1.0 - glambda;
@@ -328,7 +328,7 @@ public:
     LUE static void SetupSchemeData_2_3_3(
         TimeIntegrationAlgorithmGLMSharedPtr &phase)
     {
-        NekDouble glambda = 0.788675134594813;
+        NekDouble glambda = (3.0 + sqrt(3.0)) / 6.0;
 
         phase->m_A[0][1][1] = glambda;
         phase->m_A[0][2][1] = 1.0 - 2.0 * glambda;